+++ Assets/Prefabs/UI/ProductableUnitButton.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b9fe9189d80e855458af635f6030bd2c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
