+++ Assets/Scripts/System/UI/UnitProductPanel.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2cede7776eef5154f86f98f7a847f2c4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
