+++ Assets/Scripts/System/UI/UnitProductPanel/ProductableUnitCard.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+using UnityEngine.UI;
+using TMPro;
+
+public class ProductableUnitButton: MonoBehaviour
+{
+    public Button button;
+    public TMP_Text desc;
+    public Image icon;
+}
