+++ Assets/Scripts/System/UI/UnitProductPanel/ProductableUnitCard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 61348c99727a963458993ffc2d08c8ea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
