+++ Assets/Scripts/System/UI/UnitProductPanel/UnitProductionModuleUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d78d4d0229cc62a41a9e17a745a51eb1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
