+++ Assets/UI/Fonts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 20616df2003938f47995c9afe1c49e10
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
