+++ Assets/UI/Fonts/HongLeiXingShuJianTi-2 SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 970055a6823f7c94b813ff9963c9a545
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
